Move migrations into separate folders
